requirements.txt
python-telegram-bot==21.5
openai>=1.30.0,<1.50.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
loguru==0.7.2
tenacity==9.0.0
APScheduler==3.10.4

# Web framework and server
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Rate limiting and caching
slowapi==0.1.9
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Monitoring
prometheus-client==0.19.0

# HTTP client - Compatible with python-telegram-bot 21.5 and OpenAI
httpx>=0.27.0,<0.28.0